Add the ability to record as sudo
This functionality was originally intended to be implemented using
policykit but the added complexity of having to create a separate helper
application to be run via dbus and the added complexity to the hotspot
user of having to install the dbus service files, permissions, etc made
me lean towards kdesudo/gksudo. kdesudo/gksudo allows us to run the perf
record command as sudo without the complexity of kauth or policykit and
for our simple use case provides the functionality we need.

Issue: #11