Check correct size when reading event header
I noticed that not all context switch events where parsed
when I profiled the off-CPU values for /usr/bin/sleep. The
reason is that the code checked for the availability of
eventSize when reading an event header, instead of the
availability of sizeof(eventSize)...

Regression was introduced by f8d507, can you spot it? ;-)
Epic how I wanted to get better code by enabling warnings,
but introduced this subtle regression which isn't obviously
leading to runtime failures, even in our usual regression tests!