Add a per-CPU event timeline
When samples in the perf.data file contain the CPU id, use this
to build a list of events per CPU. This is shown in the timeline
and can be filtered like the per-thread/process timelines.

Note that you need to run perf record with `--sample-cpu` to get
the CPU ids associated with the samples. When this is not the case,
only CPU 0 will be used/available.

Fixes: #94