Do not show on-CPU time when we didn't profile off-CPU time
The values for the on-CPU times get deduced from the measured
off-CPU time. If we do not measure those, we should not try to
show any on-CPU times. This patch guards this properly, and we
won't show misleading on-CPU times in the summary page anymore
when the off-CPU time is zero.