Add support to record/profile the whole system
Fixes task #11 but we'll need to spent some more time on the
record-as-root feature... Both kdesu and gksu do not like to be
terminated. Or if they do get terminated, they don't terminate
the spawned child in turn. This makes them very much useless for
our purpose, because it would leave the `perf record` process
alive, filling up your hard disk :-/