Correctly restore the caller/callee map when unsetting filters
We used to lose the sourceMap for the caller/callee entries when
adding a filter and then unsetting all filters. The reason was
that in these cases, we never caller addCallerCaleeEvent for the
individual entries in the bottomUpResults.

Keep the original caller/callee results around and restore them
directly for this situation, to fix this bug.