Properly stop perfparser when it hangs and we don't get any input
The old code only successfully quit the perfparser process when
we actually received some data from it. If we parse a corrupted
perf.data file with bogus values, the perfparser process sometimes
just sits around not doing anything. In these cases, we couldn't
even quit the process with our old code, as no readyRead events
would arrive.

This patch refactors the whole procedure, and leverages Qt signal/
slot and event loop to kill the process directly when the user
requests it.