Add script to temporarily elevate perf privileges
Launching the script will first query the current settings for the
kernel.kptr_restrict and kernel.perf_event_paranoid sysctl values
as well as the file mode bits for /sys/kernel/debug and .../tracing.

Then the privileges are elevated, by setting kptr_restrict=0 and
perf_event_paranoid=-1 as well as remounting the debug and tracing
filesystems with mode 755.

Finally the script waits to be killed or to receive input and then
restores the previous privileges.

The plan is to use this script inplace of launching perf as root
via kdesudo, which is too brittle and not necessary for the main
use-case of hotspot (namely profiling user applications).