Never run perf as root, instead only temporarily elevate privileges
Running perf as root simply does not work out as intended. Most
notably, the graphical sudo/su applications like kdesudo, kdesu,
gksu or gksudo simply have to many limitations that make it infeasible
to run perf as root.

The current status was that hotspot ran kdesudo + perf + runuser + yourapp.
This worked when you waited for yourapp to finish, but when you interrupt
the record step we would terminate kdesudo. This in turn isn't gracefully
shutting down the perf record session, leading to garbage perf.data files
that cannot be parsed. This completely breaks the usecase of system-wide
recording, which must be run with elevated user rights and depends on
terminating perf after some time...

Instead, we now temporarily elevate the privileges, such that perf can
be used as the current non-root user. This means we set the kptr_restrict
and perf_event_paranoid sysctl knobs as well as making the debug and tracing
filesystems accessible. Then one second after we have started perf, we
undo the privilege elevation. Overall, this gives us a useable hotspot
for off-CPU profiling and system-wide recording. The only thing that
does not work is attaching to a running process from a different user.