Run `perf record` with `--aio` by default, if possible
Letting perf write the data asynchronously reduces the chances of
losing events massively. See also:

https://lore.kernel.org/patchwork/cover/1007310/

In the future, we can then also enable perf.data compression.
Obviously, we'll have to implement support for that first in
perfparser though. And it needs to actually land upstream in perf.
I'm looking forward to that a lot!