Add export menu with stack-collapse export capability
The view menu now sports an 'Export' menu. For now, only a single
export action is supported, namely exporting in the stack-collapse
format. This data can then be fed into the flamegraph.pl script
to generate interactive SVG files, if needed.

Note that multi-cost data files are supported. The user can select
what is supposed to be exported.

In the future, we can add more actions that export the data in
different formats. Most notably, we should be able to export our
flame graphs directly as static images.

Relates-To: https://github.com/KDAB/hotspot/issues/45