Add context menu to open editor for function
When activated, we try to find a source location we can open for
the symbol. If we fail, we show an error message. Otherwise we
just open the first location for the symbol we find. This is a
nice usability improvement when one just wants to open an editor
at the function. Previously, this was only possible from the source
location view in the lower right corner of the Caller/Callee view.
Now, we can basically open the editor from anywhere.

Fixes: https://github.com/KDAB/hotspot/issues/205