Unbreak tst_perfparser::testCustomCostAggregation
Do not assume the data comes from the same machine we are running
the test on. Don't use assertions and don't overwrite the expected
data file. Write to a different file and then diff that if needed.
Don't enforce UTF8 strings, write to QByteArray directly.

The expected files also seem to be wrong, I'm updating them now
based on the results from my machine.